// src/modules/comercial/produtos/roles/danny/dtos/produto-fornecedor.dto.ts
import Divergencia from '@modules/comercial/produtos/models/divergencia.model';
import { Ean } from '@modules/comercial/produtos/models/ean.model';

export default interface ProdutoFornecedorDTO {
  id: number;
  codigo_produto_fornecedor: string;
  descritivo: string;
  descritivo_pdv: string;
  classificacao_fiscal: string;
  origem: string;
  estado: number;
  marca: number;
  depto: number;
  secao: number;
  grupo: number;
  subgrupo: number;
  preco: number;
  desconto_p: number;
  pesob: number;
  pesol: number;
  altura: number;
  largura: number;
  comprimento: number;
  validade: number;
  qtde_embalagem: number;
  comprador: number;
  categoria_fiscal: string;
  produto_arius: number;
  fornecedor_id: number;
  st_compra: string;
  tipo_tributacao: string;
  ipi: number;
  icms_compra: number;
  pis_cofins: string;
  eans: Ean[];
  duns: Ean[];
  divergencias: Divergencia[];
  status: number;
  comprimento_d: number;
  largura_d: number;
  altura_d: number;
  familia: number;
}